Modify tetration method signature
As it makes no sense to have a double exponent on such operation

+++ CalculatorTest/helper/MyDoubleCalculatorClass.cs
@@ -11,7 +11,7 @@ class MyDoubleCalculatorClass : BasicArithmetic<double>
         throw new NotImplementedException();
     }
 
-    public override double Tetration(double a, double b)
+    public override double Tetration(double a, int b)
     {
         throw new NotImplementedException();
     }
